Add new fork class and eliminate validation wait loop.

+++ include/bitcoin/blockchain/interface/simple_chain.hpp
@@ -20,7 +20,7 @@
 #ifndef LIBBITCOIN_BLOCKCHAIN_SIMPLE_CHAIN_HPP
 #define LIBBITCOIN_BLOCKCHAIN_SIMPLE_CHAIN_HPP
 
-#include <cstdint>
+#include <cstddef>
 #include <bitcoin/bitcoin.hpp>
 #include <bitcoin/blockchain/define.hpp>
 
@@ -38,67 +38,65 @@ public:
     // ------------------------------------------------------------------------
 
     /// Return the first and last gaps in the blockchain, or false if none.
-    virtual bool get_gap_range(uint64_t& out_first,
-        uint64_t& out_last) const = 0;
+    virtual bool get_gap_range(size_t& out_first, size_t& out_last) const = 0;
 
     /// Return the next chain gap at or after the specified start height.
-    virtual bool get_next_gap(uint64_t& out_height,
-        uint64_t start_height) const = 0;
+    virtual bool get_next_gap(size_t& out_height,
+        size_t start_height) const = 0;
 
     /// Get a determination of whether the block hash exists in the store.
     virtual bool get_exists(const hash_digest& block_hash) const = 0;
 
     /// Get the difficulty of the branch starting at the given height.
     virtual bool get_difficulty(hash_number& out_difficulty,
-        uint64_t from_height) const = 0;
+        size_t from_height) const = 0;
 
     /// Get the header of the block at the given height.
     virtual bool get_header(chain::header& out_header,
-        uint64_t height) const = 0;
+        size_t height) const = 0;
 
     /// Get the height of the block with the given hash.
-    virtual bool get_height(uint64_t& out_height,
+    virtual bool get_height(size_t& out_height,
         const hash_digest& block_hash) const = 0;
 
     /// Get the bits of the block with the given height.
-    virtual bool get_bits(uint32_t& out_bits,
-        const uint64_t& height) const = 0;
+    virtual bool get_bits(uint32_t& out_bits, const size_t& height) const = 0;
 
     /// Get the timestamp of the block with the given height.
     virtual bool get_timestamp(uint32_t& out_timestamp,
-        const uint64_t& height) const = 0;
+        const size_t& height) const = 0;
 
     /// Get the version of the block with the given height.
     virtual bool get_version(uint32_t& out_version,
-        const uint64_t& height) const = 0;
+        const size_t& height) const = 0;
 
     /// Get height of latest block.
-    virtual bool get_last_height(uint64_t& out_height) const = 0;
+    virtual bool get_last_height(size_t& out_height) const = 0;
 
     /// Get the hash digest of the transaction of the outpoint.
     virtual bool get_transaction_hash(hash_digest& out_hash,
         const chain::output_point& outpoint) const = 0;
 
     /// Get the block height of the transaction given its hash.
-    virtual bool get_transaction_height(uint64_t& out_block_height,
+    virtual bool get_transaction_height(size_t& out_block_height,
         const hash_digest& transaction_hash) const = 0;
 
     /// Get the transaction of the given hash and its block height.
-    virtual transaction_ptr get_transaction(uint64_t& out_block_height,
+    virtual transaction_ptr get_transaction(size_t& out_block_height,
         const hash_digest& transaction_hash) const = 0;
 
     // Setters.
     // ------------------------------------------------------------------------
 
     /// Insert a block to the blockchain, height is checked for existence.
-    virtual bool insert(block_const_ptr block, uint64_t height) = 0;
+    virtual bool insert(block_const_ptr block, size_t height) = 0;
 
     /// Append the block to the top of the chain, height is validated.
-    virtual bool push(block_const_ptr block, uint64_t height) = 0;
+    virtual bool push(block_const_ptr block, size_t height) = 0;
 
     /// Remove blocks at or above the given height, returning them in order.
     virtual bool pop_from(block_const_ptr_list& out_blocks,
-        uint64_t height) = 0;
+        size_t height) = 0;
 };
 
 } // namespace blockchain
